separation of lib/ into common/, compress/, decompress/, dictBuilder/, legacy/
