+++ lib/common/bitstream.h
