+++ lib/common/error_private.h
