+++ lib/common/mem.h
