+++ lib/common/zbuff.h
