+++ lib/compress/zstd_compress.c
@@ -55,7 +55,7 @@
 #include <string.h>   /* memset */
 #include "mem.h"
 #include "fse_static.h"
-#include "huff0_static.h"
+#include "huf_static.h"
 #include "zstd_internal.h"
