+++ lib/dictBuilder/zdict_static.h
