+++ lib/legacy/zstd_v04.c
@@ -2004,11 +2004,6 @@ extern "C" {
 #endif
 
 
-/* ****************************************
-*  Dependency
-******************************************/
-#include "huff0.h"
-
 
 /* ****************************************
 *  Static allocation macros
